requirements.txt
absl-py==0.9.0
albumentations==0.4.5
astor==0.8.1
astroid==2.4.0
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.2
backcall==0.1.0
bleach==1.5.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defusedxml==0.6.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@e1252f9aab632a23c40787caf700745a2e285a58
easydict==1.7
entrypoints==0.3
enum34==1.1.10
fire==0.3.1
future==0.18.2
fvcore==0.1.1.post20200716
gast==0.3.3
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.28.1
h5py==2.10.0
html5lib==0.9999999
idna==2.9
imageio==2.8.0
imgaug==0.2.6
imgviz==1.1.0
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
json-tricks==3.15.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
labelme==4.4.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
munch==2.5.0
mxnet-cu102==1.6.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
ninja==1.10.0
notebook==6.0.3
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
portalocker==1.7.0
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pycodestyle==2.5.0
pydot==1.4.1
Pygments==2.6.1
pylint==2.5.0
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
rsa==4.0
scikit-image==0.17.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.6.4
six==1.14.0
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.7.4
toml==0.10.0
torch==1.5.0
torchsummary==1.5.1
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.7
zipp==3.1.0
